requirements.txt
datasets==2.13.1
transformers==4.30.2
pillow>=9.4.0
huggingface_hub>=0.15.1